requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic[email-validator]==2.5.2
python-dotenv==1.0.0

# Dashboard
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas>=2.2.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
hypothesis==6.92.1

# Development
black==23.11.0
isort==5.12.0
